test: create entities without any relations defined in the model
